+++ udpcap/src/udpcap_socket.cpp
@@ -21,6 +21,10 @@
 
 #include "udpcap_socket_private.h"
 
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+
 namespace Udpcap
 {
   UdpcapSocket::UdpcapSocket()
@@ -45,10 +49,10 @@ namespace Udpcap
 
   bool              UdpcapSocket::hasPendingDatagrams        () const                                                { return udpcap_socket_private_->hasPendingDatagrams(); }
 
-  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, unsigned long timeout_ms, HostAddress* source_address, uint16_t* source_port, Udpcap::Error& error) { return udpcap_socket_private_->receiveDatagram(data, max_len, timeout_ms, source_address, source_port, error); }
-  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, unsigned long timeout_ms, Udpcap::Error& error)                                                     { return udpcap_socket_private_->receiveDatagram(data, max_len, timeout_ms, nullptr, nullptr, error); }
-  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, Udpcap::Error& error)                                                                               { return udpcap_socket_private_->receiveDatagram(data, max_len, -1, nullptr, nullptr, error); }
-  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, HostAddress* source_address, uint16_t* source_port, Udpcap::Error& error)                           { return udpcap_socket_private_->receiveDatagram(data, max_len, -1, source_address, source_port, error); }
+  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, long long timeout_ms, HostAddress* source_address, uint16_t* source_port, Udpcap::Error& error) { return udpcap_socket_private_->receiveDatagram(data, max_len, timeout_ms, source_address, source_port, error); }
+  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, long long timeout_ms, Udpcap::Error& error)                                                     { return udpcap_socket_private_->receiveDatagram(data, max_len, timeout_ms, nullptr, nullptr, error); }
+  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, Udpcap::Error& error)                                                                           { return udpcap_socket_private_->receiveDatagram(data, max_len, -1, nullptr, nullptr, error); }
+  size_t            UdpcapSocket::receiveDatagram(char* data, size_t max_len, HostAddress* source_address, uint16_t* source_port, Udpcap::Error& error)                       { return udpcap_socket_private_->receiveDatagram(data, max_len, -1, source_address, source_port, error); }
 
   bool              UdpcapSocket::joinMulticastGroup         (const HostAddress& group_address)                      { return udpcap_socket_private_->joinMulticastGroup(group_address); }
   bool              UdpcapSocket::leaveMulticastGroup        (const HostAddress& group_address)                      { return udpcap_socket_private_->leaveMulticastGroup(group_address); }
